Add a python flask application for Github API
This application will keep track of Github repository pull requests for
the purpose of allowing one to view the diff between subsequent force
push attempts to the branch.  Github does not currently have a way to
view the old head of the branch prior to a force push.  This means that
if someone addresses comments on a PR they have opened and force pushes
the changes, the reviewers are unable to get a diff of what actually
changed.

This application will keep track of the branches corresponding to open
pull requests and will post a comment on the Github pull request page
whenever someone force pushes the branch.  The comment will contain
links to the diff of the overall code, the diff of the commit logs (with
or without associated commit diffs), and diff series where one can view
up to 3 previous versions of the branch code or logs.

This also will check the branch for fixup! or squash! commits and mark
those commit status as "failed" to indicate that branches with such
commits should not be merged.