Update commit validation to check more cases
_check_for_fixup_or_squash_commits was written to check whether there
were any fixup! or squash! commits in the branch that's currently
associated with a pull request.  There are other checks we can make for
each commit besides that.  Add methods to check the author and committer
information, the commit message format itself, and whether the
associated diff introduces any whitespace issues (e.g., trailing
whitespace or extra blank lines at the end of the file).  Call the
method to make those checks instead of using
_check_for_fixup_or_squash_commits, and use the list of errors it
returns (if any) for a given commit as part of the status message posted
to the Github status API.  Remove _check_for_fixup_or_squash_commits
since it's no longer used and has be superceded by the new methods
added.