Fix issue where push did not trigger status check
On occasion, Github will send webhook events before updating the actual
repo with the updated branch information.  This lead to outdated
information being retrieved by the git ls-remote command which lead to
the issue where the branch could not be associated with a pull request.

To fix this, a call to time.sleep(1) was added before calling git ls-remote.

Another issue was noted when marking the state of commits in a branch.
If the head commit did not have a failure state, then Github would still
present the green "Merge pull request" button in the user interface even
if other commits in the same branch were marked as failures.  To fix
this, we check whether any commit in the branch was marked with a
failure state.  If so, the we mark the head commit of the branch with a
failure state to prevent the green "Merge pull request" button from
showing up in the user interface.