Fix bug for head commit status on PR open event
Github will consider a PR mergable even if some of the commits in the
branch are marked as failure.  To get around this, we check if any
commit has been marked as failure and then set the head commit status to
the same value.  This commit fixes an issue where this wasn't happening
when opening a PR (though it did work when pushing additional commits on
an open PR.

Also add some time.sleep(1) calls before git fetch ooperations to
mitigate any possible race conditions.