Reduce rebase comment length
The original branch has been rebased comment format was too verbose and
had unnecessary links.  The comment length was reduced by reducing the
number of lines and removing links to diffs (head to head for branch and
commit log) that weren't really useful.

The bug that resulted in the first line being indented was fixed.  Also,
a docstring explaining the various diffs was added to the method.