Make commit message body presence check mandatory
In the vast majority of cases, a commit message needs a body to explain
what was done and why it was done.  So, requiring a body for the commit
message makes more sense than making it optional.

This change removes the separator check condition before checking for
the presence of the commit message body and whether each line in it is
less than or equal to 72 characters in length.