Add sleep calls before git fetch invocations
As mentioned in:

    1ec2346 "Fix issue where push did not trigger status check"

and

    3a6f53d "Fix bug for head commit status on PR open event"

Github will sometimes send a request to the webhook before updating
their repository refs.  This can result in issues when creating local
references based on the remote reference state which can result in
either no diff or the wrong diff being shown when calling one of the git
diff endpoints.  The one second delay introduced before running a git
fetch should fix this issue in the vast majority of cases.