Add close fd call after running git diff --check
One of the checks run on a commit is to check whether it will introduce
any whitespace issues like trailing whitespace, blank lines at the end
of the file, or blank lines that contain whitespace.  This is done by
running git diff --check via subprocess.  This command writes to the
stdout and stderr file descriptors, so they are provided to the child
process by opening a temporary file and providing its file descriptor to
the subprocess call in the stdout and stderr named parameters.  The
close_fds named parameter is also set to true.

This means that the child process which runs git diff --check will close
that file descriptor, but there was no corresponding close call in the
parent process.  Since the flaks application is a long running process,
this eventually lead to the condition where the process would run out of
unused file descriptors, which in turn would prevent any calls to
subprocess from working (due to the OSError: [Errno 24] Too many open
files) exception that was raised.

This adds an os.close() call on the file descriptor on that temporary
file in the parent process which should prevent this error from
occurring (since the file descriptor will now be closed in both the
parent and child processes).