Use separate Github statuses for commit checks
The status endpoint in the Github API takes a description parameter, which
according to the documentation at

https://developer.github.com/v3/repos/statuses/#create-a-status

can be set to "[a] short description of the status".

If a commit has multiple issues with it, the original way of generating the
description string by concatenating all the error strings would result in a
string that was too long.  This resulted in a failure when attempting a HTTP
post request to the endpoint noted above.  In order to get around this issue,
update the implementation to use a separate context and description for each
issue found with a commit.  Using a different context for each check allows for
the UI to display all issues with a particular commit.

Using the same context string for each status post was attempted, but it turned
out that only the most recent status set would be displayed in the UI (even
though the documentation states that one can set up to 1000 statuses per
commit).