Fix rebase off-by-one message bug
When a branch associated with a pull request is rebased and force-pushed
to github, the rebase check will post a comment noting that the branch
has been rebased.  The message begins with a sentence stating the total
number of times the branch has been rebased since the PR was opened.  An
off-by-one bug resulted in the message showing a number one less than it
should have been.  For example, on the first rebase, the message would
start with:

    Branch rebased 0 time(s), most recently by ...

instead of saying 1 time(s).

Fix the bug by adding one to the variable used to show that number.