Remove unused code in _parse_commit_log method
There was some code after the final return statement in
_parse_commit_log that may have been left over from copying/pasting when
refactoring the code base prior to the first commit.  Since this code
would never be run and the code is used elsewhere in the code base to
handle setting the pull request branch head commit status based on the
branch state, it was removed.