Add --reverse parameter to git log command
In _parse_commit_log, git log --format <commit_range> is called to get
log output.  In order to prepare for the addition of the git diff churn
check (via the -S parameter of git log), we need to reverse the commit
order.  This will allow the diff churn check to first run on the earliest
commit in the branch instead of the latest commit.

We want to do this because that check looks at later commits in the
branch to see whether changes intoduced by those commits alter changes
made in the commit we're checking.  Starting from the earliest commit
will allow for flagging later commits if they fail the check described
earlier.  If those commits are already flagged, then there's no need to
directly check them again (which saves time).

If we started from the latest commit instead, then we would have to
repeat the checks for each parent commit.