Add _parse_diff method
This method will be used as part of the diff churn check that will be
added to the gitbot in a future commit.  The check will involve running
git log -S'<diff_line_to_check>  <commit_sha1>.. and checking for any
<commit_sha1> values that show up in the result.  This will end up
flagging commits that update code that was updated in the commit that is
being checked (much like a fixup! commit normally does).