Fix bug introduced by git log --reverse prameter
The _parse_commit_log method uses a state machine to parse the output of
git log --format=full. The original implementation has a bug that will
allow multiple if statements to execute if the last commit in the log
does not have a commit message body.  This resulted in the separator and
body strings to get populated with the value of title.  This resulted in
a commit without a message body to not get flagged as such.

Add break statements to those if blocks that will execute if we have
reached the end of the git log output.  This should prevent separator
and body from getting assigned the value of title.