Add _check_diff_add_delete method
This method will run

    git log -S'<diff_line>' <commit_sha1>..

for each unique line in the diff (including whitespace) for the
commitish value that's passed in as a parameter.  This will be used in a
future commit to determine if any subsequent commits in the branch
affect lines that were updated by the commit in question.  This will
allow for detecting churn in the overall diff of a branch (meaning that
subsequent commits undo changes made in an earlier commit).