Fix git commit address check
The address check normalizes the author and committer display name and
email address domain by calling the lower method on the string it parses
from the commit message.  The lower method is a method that can be
called on a string object.  The original implementation called it as a
global method which isn't valid code.  The bug was triggered when a
commit was pushed on a repository where the author and committer email
address didn't have a domain that matched one of the valid domains
specified in the config file.

Update the code to call the lower method as a method of the string for
the author and committer root, real-name, and valid-domain checks to fix
the bug.