Add calls to check commit add_delete and move
We want to be able to determine whether later commits in a branch modify
code that was changed in an earlier commit in the same branch.  This can
happen in several ways:

Lines added in a given commit are either:
1. Removed in a subsequent commit
2. Duplidated in a subsequent commit
3. Moved elsewhere in a subsequent commit

Lines removed in a given commit are either:
1. Restored in a subsequent commit
2. Identical lines are removed in a subsequent commit
3. Moved elsewhere in a subsequent commit

This usually means that the changes should be in a single commit or
should be in seqprate branches.